Add addNewTodo clearAllCompletedTodos  editTodo and onDeleteTodo functions  in add-remove.js file
